[BACKLOG-4490] Changed Karaf version 3.0.3. Added the org.apache.util.maven to karaf classpath, and change start level for the Pentaho deployment dependencies to start before the boot features.
Changed Karaf version from 2.3.5 to 3.0.3, as well as, the karaf
features.xml version. Changed the features.xml and startup.properties to
reflect the karaf version modification. 
Add the dependency org.apache.util and use it in the standalone-assembly
to include its .jar file inside the lib folder of karaf. This way, the
org.apache.util.maven is included in the karaf classpath.
In startup.properties, I changed the start level for the Pentaho
deployment dependencies to start before the boot features start because
the protocol pentaho-platform-plugin-mvn used in the
pentaho-data-profiling-ce-standalone feature depends on the
pentaho-platform-plugin-deployer bundle.